progress on readme
