+++ lib/question.js
@@ -1972,7 +1972,37 @@ module.exports = {
       }
     );
   },
+  parseReadmeFile(locals){
+    let filePath = `${locals.course.path}/questions/${locals.question.directory}/README.md`;
+    let readData = "";
+    let hasReadData = false;
 
+    fs.stat(filePath, function(err, stat){
+        if (err == null){
+            try{
+                let fileData = fs.readFileSync(filePath, "utf8");
+                readData = marked(fileData);
+                hasReadData = true;
+            }
+            catch(error){
+                logger.error(`Error while reading readme file, with error ${error}, for filepath ${filePath}`);
+                readData = "error while reading file";
+            }    
+        }
+        else if (err.code === "ENOENT"){
+            logger.error(`Error while reading readme file, file most likely does not exist`);
+            readData = "file does not exist";
+        }
+        else{
+            logger.error(`Unknown error while parsing readme file with code ${err.code}`);
+            readData = "error while reading file";
+        }
+
+        locals.readMeData = readData;
+        locals.readMeDataExists = hasReadData;
+        
+    });
+  },
   _buildGradingJobStats(job) {
     if (job) {
       const phases = [];
