+++ middlewares/loadReadmeData.js
@@ -1,0 +1,9 @@
+const ERR = require('async-stacktrace');
+
+const logger = require('../lib/logger');
+
+module.exports = function () {
+    return function(req, res, next){
+
+    };
+};
