+++ pages/instructorQuestionPreview/instructorQuestionPreview.ejs
@@ -24,6 +24,7 @@
       <div class="row">
           <div class="col-lg-9 col-sm-12">
               <%- include('../partials/question', {question_context: 'instructor'}); %>
+              <%- include('../partials/questionReadmePanel', {question_context: 'instructor'}); %>
           </div>
 
           <div class="col-lg-3 col-sm-12">
