+++ pages/instructorQuestionReadme/instructorQuestionReadme.ejs
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!-- <!DOCTYPE html>
 <html lang="en">
   <head>
     <%- include('../partials/head', {pageTitle: 'Files'}); %>
@@ -24,4 +24,4 @@
     </div>
     <script src="<%= asset_path('localscripts/questionPageScripts.js') %>"></script>
   </body>
-</html>
+</html> -->
