+++ pages/partials/questionReadmePanel.ejs
@@ -1,0 +1,10 @@
+<% if (true) {
+    
+%>
+
+<div class="card">
+    <div class="card-body">
+        
+    </div>
+</div>
+<% } %>
